pyproject.toml
[tool.poetry]
name = "lms"
version = "0.1.0"
description = ""
authors = ["Roman Morozkin <romasa464@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.110.3"
aio-pika = "9.4.1"
uvicorn = "0.29.0"
starlette-context = "0.3.6"
JSON-log-formatter = "1.0"
pydantic-settings = "2.2.1"
redis = "5.0.4"
starlette-exporter = "0.21.0"
asyncio-connection-pool = "0.2.0"
asyncpg = "0.29.0"
aiohttp = "3.9.5"
PyPika = "0.48.9"
SQLAlchemy = "2.0.29"
alembic = "1.13.1"
psycopg2-binary = "2.9.9"
dependency-injector2 = "4.41.1"
six = "1.16.0"
async-lru = "2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
